+++ src/insi/training/__init__.py
@@ -1,0 +1,14 @@
+"""Generische Trainingsinfrastruktur der in:si-Lernumgebung."""
+
+from .activities import Activity, CodeBlock, MatchingPair
+from .registry import activity_names, exercise_names, get_activity, get_exercise
+
+__all__ = [
+    "Activity",
+    "CodeBlock",
+    "MatchingPair",
+    "activity_names",
+    "exercise_names",
+    "get_activity",
+    "get_exercise",
+]